[refactor]: 拆分 BoardPanel.tsx 到 XxxJSX.tsx 多个文档中
