+++ src/Pages/Game/CapturedPiecePanel.tsx
@@ -1,6 +1,7 @@
-import {map} from "./BoardPanel";
+import {map} from "./Board/BoardJSX";
 import {CapturedPieceData, GridData} from "../../Game/GetData";
-import {Grid} from "./BoardPanel";
+
+import {GridJSX} from "./Board/GridJSX";
 
 
 export function CapturedPiecePanel({capturedPieceData} : {capturedPieceData: CapturedPieceData}){
@@ -8,7 +9,7 @@ export function CapturedPiecePanel({capturedPieceData} : {capturedPieceData: Cap
         {
             map(
                 capturedPieceData.pieces.length, (x) =>
-                    <Grid
+                    <GridJSX
                         grid = {new GridData(false, capturedPieceData.pieces[x].pieceData, 0)}
                         handleClick = {() => {}}
                     />
