+++ src/Pages/Game/GamePanel.tsx
@@ -1,4 +1,4 @@
-import {BoardPanel} from "./BoardPanel";
+import {BoardJSX} from "./Board/BoardJSX";
 import {Game} from "../../Game/Game";
 import {GetData} from "../../Game/GetData";
 import {useState} from "react";
@@ -14,13 +14,13 @@ export function GamePanel({game}: {game: Game}) {
 
     return (
         <div id="Game" className="Game">
-            <BoardPanel boardData={GetData.GetBoardData(game, game.board)}
-                        handleClick={(x, y) => {
+            <BoardJSX boardData={GetData.GetBoardData(game, game.board)}
+                      handleClick={(x, y) => {
                             game.board.handleClick(x, y, game.players.current)
                             setCount(count+1);
                         }
                         }
-                        rowNoType="number" columnNoType="chinese"/>
+                      rowNoType="number" columnNoType="chinese"/>
         </div>
     )
 }
